requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
python-dotenv==1.0.0
openai==1.78.1
pandas==2.2.1
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0
